Move completely except test code

+++ contactscommon/src/main/java/com/android/contacts/common/ContactsUtils.java
@@ -98,8 +98,8 @@ public class ContactsUtils {
     }
 
 
-    public static final long USER_TYPE_CURRENT = 0;
-    public static final long USER_TYPE_WORK = 1;
+    public static final int USER_TYPE_CURRENT = 0;
+    public static final int USER_TYPE_WORK = 1;
 
     /**
      * UserType indicates the user type of the contact. If the contact is from Work User (Work
